:recycle: Rename component VotingV2 to Voting
