+++ src/components/voteUrlProcessorV2/voteUrlProcessor.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { parseSearchParams } from '../../utils/searchParams';
 import routes from '../../constants/routes';
 
-import styles from '../votingV2/votingV2.css';
+import styles from '../voting/voting.css';
 
 export default class VoteUrlProcessor extends React.Component {
   constructor(props) {
