+++ src/components/voting/voting.css
