+++ src/components/voting/votingResult.css
