+++ src/components/voting/votingResult.js
