+++ src/constants/routes.js
@@ -9,7 +9,7 @@ import SearchResult from '../components/searchResult';
 import TransactionDashboard from '../components/transactionDashboard';
 import AccountTransactions from '../components/accountTransactions';
 import Delegates from '../components/delegatesV2';
-import Voting from '../components/votingV2';
+import Voting from '../components/voting';
 import SingleTransaction from '../components/singleTransactionV2';
 import HwWalletLogin from '../components/hwWalletLogin';
 // import NotFound from '../components/notFound';
